\docType{methods}
\name{get_pathrow}
\alias{get_pathrow}
\title{Get WRS-2 path/row for a given spatial object}
\usage{
get_pathrow(x)
}
\arguments{
  \item{x}{a spatial object}
}
\value{
list with path and row as integers
}
\description{
Get WRS-2 path/row for a given spatial object
}
\examples{
#TODO: add examples
}

